requirements.txt
antlr4-python3-runtime==4.13.2
antlr4-tools==0.2.2
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.2
gunicorn==23.0.0
install-jdk==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
Werkzeug==3.1.4
